// workers/src/routes/phieu/types.ts
export type PhieuScopeUser = { username: string; role: string };

export type ResultScope = {
  result_id: string;
  student_id: string;
  student_name: string;
  class_id: string;
  teacher_username: string;
  mon_hoc: string;
  ten_bai_tap: string;
  ngay_lam_bai: string;
  tong_cau: number;
  so_cau_dung: number;
  so_cau_sai: number;
  diem_so: number;
};